requirements.txt
fastapi
uvicorn
motor
python-dotenv
bcrypt
PyJWT
pydantic
python-multipart
passlib
